mamangementsellTableIdentityProblemOndb

+++ Entity/Management.cs
@@ -5,8 +5,8 @@ public class Management: IEntity
     public int Id { get; set; }
     
     public int Type { get; set; }
-    public float AdminCost { get; set; }
-    public float UserCost { get; set; }
+    public double AdminCost { get; set; }
+    public double UserCost { get; set; }
     public DateTime UserStartDate { get; set; }
     public DateTime UserEndDate { get; set; }
     public int UserWorkedDayCount { get; set; }
